requirements.txt
slackeventsapi==2.2.1
Flask==1.1.2
requests==2.24.0
slackclient==2.9.3
